+++ src/funding/card/discover/config.js
@@ -3,9 +3,10 @@
 import { DiscoverLogo } from '@paypal/sdk-logos/src';
 
 import type { CardConfig } from '../../common';
+import { enableLogoCDNExperiment } from '../../../lib/getLogoCDNExperiment';
 
 export function getDiscoverConfig() : CardConfig {
     return {
-        Label: DiscoverLogo
+        Label: () => enableLogoCDNExperiment(DiscoverLogo)
     };
 }