+++ src/funding/paylater/config.jsx
@@ -9,6 +9,7 @@ import { PPLogo, LOGO_COLOR } from '@paypal/sdk-logos/src';
 import { BUTTON_COLOR, BUTTON_LAYOUT, DEFAULT } from '../../constants';
 import { DEFAULT_FUNDING_CONFIG, type FundingSourceConfig } from '../common';
 import { Text, Space } from '../../ui/text';
+import { enableLogoCDNExperiment } from '../../lib/getLogoCDNExperiment';
 
 import css from './style.scoped.scss';
 
@@ -77,7 +78,7 @@ export function getPaylaterConfig() : FundingSourceConfig {
         Logo: ({ logoColor, nonce, fundingEligibility }) => {
             return (
                 <Style css={ css } nonce={ nonce }>
-                    <PPLogo logoColor={ logoColor } />
+                    {enableLogoCDNExperiment(PPLogo, { logoColor })}
                     <Space />
                     <Text>{ getLabelText(fundingEligibility) || 'Pay Later' }</Text>
                 </Style>