+++ src/funding/venmo/template.jsx
@@ -9,6 +9,7 @@ import {
     BasicLabel
 } from '../common';
 import { Text, Space } from '../../ui/text';
+import { enableLogoCDNExperiment } from '../../lib/getLogoCDNExperiment';
 
 import css from './style.scoped.scss';
  
@@ -17,7 +18,7 @@ export function AppLabel(opts : LabelOptions) : ChildType {
 
     const AppLogo : ChildType = (
         <Style css={ css }>
-            <VenmoLogo logoColor={ logoColor } />
+            {enableLogoCDNExperiment(VenmoLogo, { logoColor })}
             <Text className={ [ 'app-label' ] }>
                 App
             </Text>
@@ -38,7 +39,7 @@ export function Label(opts : LabelOptions) : ChildType {
 export function  WalletLabel({ ...props } : WalletLabelOptions) : ChildType {
     const { instrument, logoColor } = props;
     let label;
-    const logo =  <VenmoLogo logoColor={ logoColor } />;
+    const logo =  enableLogoCDNExperiment(VenmoLogo, { logoColor });
 
     if (instrument && instrument.label) {
         label = instrument.label;