+++ test/percy/.eslintrc.js
@@ -1,0 +1,11 @@
+/* @flow */
+
+// eslint-disable-next-line import/no-commonjs
+module.exports = {
+  rules: {
+    "no-restricted-globals": ["error"],
+    "promise/no-native": "off",
+    "flowtype/require-valid-file-annotation": "off",
+    "flowtype/require-return-type": "off",
+  },
+};