+++ test/percy/content.js
@@ -1,0 +1,17 @@
+export const testContent = {
+  instantlyPayWith: "Pay instantly with",
+  poweredBy: "Powered by PayPal",
+  chooseCardOrShipping: "Choose card or shipping",
+  useDifferentAccount: "Use different account",
+  deleteVaultedAccount: "Forget this account",
+  deleteVaultedCard: "Forget this card",
+  chooseCard: "Choose card",
+  payWithDebitOrCreditCard: "Debit or Credit Card",
+  payWithDifferentAccount: "Pay with a different account",
+  payWithDifferentMethod: "Pay with a different funding method",
+  payNow: "Pay Now",
+  payWith: "Pay with",
+  credit: "Credit",
+  balance: "Balance",
+  payPalBalance: "PayPal balance",
+};